+++ packages/react-native-worklets/plugin/index.js
@@ -1077,6 +1077,9 @@ var require_globals = __commonJS({
       // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/arguments
       "require",
       "fetch",
+      "Headers",
+      "Request",
+      "Response",
       "XMLHttpRequest",
       "Blob",
       "FileReader",